+++ internal/backends/selector/api.rs
@@ -47,8 +47,8 @@ pub struct BackendSelector {
     #[cfg(feature = "unstable-winit-030")]
     winit_custom_application_handler:
         Option<Box<dyn i_slint_backend_winit::CustomApplicationHandler>>,
-    #[cfg(all(target_os = "linux", feature = "unstable-input-09"))]
-    input_090_event_hook: Option<Box<dyn Fn(&input::Event) -> bool>>,
+    #[cfg(all(target_os = "linux", feature = "unstable-libinput-09"))]
+    libinput_event_hook: Option<Box<dyn Fn(&input::Event) -> bool>>,
 }
 
 impl BackendSelector {
@@ -203,15 +203,15 @@ impl BackendSelector {
     /// The provided hook is invoked for every event received. If the function returns true, the event is
     /// not dispatched further.
     ///
-    /// *Note*: This function is behind the [`unstable-input-09` feature flag](slint:rust:slint/docs/cargo_features/#backends)
+    /// *Note*: This function is behind the [`unstable-libinput-09` feature flag](slint:rust:slint/docs/cargo_features/#backends)
     ///         and may be removed or changed in future minor releases, as new major Winit releases become available.
     #[must_use]
-    #[cfg(all(target_os = "linux", feature = "unstable-input-09"))]
-    pub fn with_input_090_event_hook(
+    #[cfg(all(target_os = "linux", feature = "unstable-libinput-09"))]
+    pub fn with_libinput_event_hook(
         mut self,
         event_hook: impl Fn(&input::Event) -> bool + 'static,
     ) -> Self {
-        self.input_090_event_hook = Some(Box::new(event_hook));
+        self.libinput_event_hook = Some(Box::new(event_hook));
         self
     }
 
@@ -285,9 +285,9 @@ impl BackendSelector {
                     builder = builder.with_renderer_name(renderer_name.into());
                 }
 
-                #[cfg(all(target_os = "linux", feature = "unstable-input-09"))]
-                if let Some(event_hook) = self.input_090_event_hook.take() {
-                    builder = builder.with_input_event_hook(event_hook);
+                #[cfg(all(target_os = "linux", feature = "unstable-libinput-09"))]
+                if let Some(event_hook) = self.libinput_event_hook.take() {
+                    builder = builder.with_libinput_event_hook(event_hook);
                 }
 
                 Box::new(builder.build()?)